+++ src/components/PrCode.vue
@@ -3,8 +3,6 @@
      VAG group (VW/Audi/Skoda/Seat/Porsche/Bentley/Lamborghini/Ducati/Cupra/Scania/MAN) manufacturer PR-Code
 */
 
-import type { PropType } from 'vue';
-
 const props = defineProps({
   prcode: {
     type: String,