+++ src/components/Product/ProductModels.vue
@@ -11,8 +11,16 @@ const props = defineProps({
 </script>
 
 <template>
-  <div inline-flex flex-wrap max-w-max>
-    <span v-for="(modelId, index) in modelIds" :key="index" class="product-model block">
+  <div
+    inline-flex
+    flex-wrap
+    max-w-max
+  >
+    <span
+      v-for="(modelId, index) in props.modelIds"
+      :key="index"
+      class="product-model block"
+    >
       <ProductModel :id="modelId" />
     </span>
   </div>