+++ src/components/Table.vue
@@ -14,14 +14,26 @@ const capitalizeFirstLetter = (text: string) => {
   <table class="table-auto text-left border bg-white shadow-md">
     <thead class="bg-gray-500 text-white">
       <tr class="border">
-        <th class="px-4 py-2 font-semibold" v-for="(thead, index) in theads" :key="index">
+        <th
+          v-for="(thead, index) in theads"
+          :key="index"
+          class="px-4 py-2 font-semibold"
+        >
           {{ capitalizeFirstLetter(thead) }}
         </th>
       </tr>
     </thead>
     <tbody>
-      <tr class="border" v-for="row in props.data" :key="row">
-        <td class="px-4 py-2" v-for="key in Object.keys(row)" :key="key">
+      <tr
+        v-for="(row, index) in props.data"
+        :key="index"
+        class="border"
+      >
+        <td
+          v-for="key in Object.keys(row)"
+          :key="key"
+          class="px-4 py-2"
+        >
           {{ row[key] }}
         </td>
       </tr>