+++ CHANGELOG.md
@@ -1,3 +1,77 @@
+## [1.1.17](https://github.com/polo-blue/sds/compare/v1.1.16...v1.1.17) (2025-09-23)
+
+## [1.1.16](https://github.com/polo-blue/sds/compare/v1.1.15...v1.1.16) (2025-09-23)
+
+### Bug Fixes
+
+* apply prettier formatting to resolve CI formatting issues ([a95d274](https://github.com/polo-blue/sds/commit/a95d2740358fa6941055f8957a40936a3175fa3e))
+
+## [1.1.15](https://github.com/polo-blue/sds/compare/v1.1.14...v1.1.15) (2025-09-23)
+
+### Bug Fixes
+
+* resolve majority of ESLint warnings ([54c4b7f](https://github.com/polo-blue/sds/commit/54c4b7f611301523ef61359cd62e818026033549))
+
+## [1.1.14](https://github.com/polo-blue/sds/compare/v1.1.13...v1.1.14) (2025-09-23)
+
+### Bug Fixes
+
+* use props.label in MainInput.vue template to resolve unused props error ([abd1e7b](https://github.com/polo-blue/sds/commit/abd1e7b724873b005d0cf5274af4aa189653c94a))
+
+## [1.1.13](https://github.com/polo-blue/sds/compare/v1.1.12...v1.1.13) (2025-09-23)
+
+### Bug Fixes
+
+* apply prettier formatting to ProductDetails.vue ([4b4c508](https://github.com/polo-blue/sds/commit/4b4c508ea86c050bc8af398484469f172cdea5b8))
+
+## [1.1.12](https://github.com/polo-blue/sds/compare/v1.1.11...v1.1.12) (2025-09-23)
+
+### Bug Fixes
+
+* restore normal console usage and add browser globals to ESLint config ([b8863af](https://github.com/polo-blue/sds/commit/b8863af8ec11266c8d612ea28129423a51d0a2d7))
+
+## [1.1.11](https://github.com/polo-blue/sds/compare/v1.1.10...v1.1.11) (2025-09-23)
+
+### Bug Fixes
+
+* resolve remaining ESLint errors ([f95095f](https://github.com/polo-blue/sds/commit/f95095feea0d8e72f41dea10660d942d3b6a02f7))
+
+## [1.1.10](https://github.com/polo-blue/sds/compare/v1.1.9...v1.1.10) (2025-09-23)
+
+### Bug Fixes
+
+* apply prettier formatting to PrCode.vue ([7de045c](https://github.com/polo-blue/sds/commit/7de045c0749153c1e1301e2223cbbcd18483cd23))
+
+## [1.1.9](https://github.com/polo-blue/sds/compare/v1.1.8...v1.1.9) (2025-09-23)
+
+### Bug Fixes
+
+* resolve critical ESLint errors in Vue components ([b95d043](https://github.com/polo-blue/sds/commit/b95d04384c9c676c2caeab1118c5ac7934498340))
+
+## [1.1.8](https://github.com/polo-blue/sds/compare/v1.1.7...v1.1.8) (2025-09-23)
+
+### Bug Fixes
+
+* update ESLint configuration to properly parse Vue files ([0eb941a](https://github.com/polo-blue/sds/commit/0eb941a37002c80753f1324c30873888cfe62f70))
+
+## [1.1.7](https://github.com/polo-blue/sds/compare/v1.1.6...v1.1.7) (2025-09-23)
+
+### Bug Fixes
+
+* remove unnecessary escape characters in text.ts ([f1ae714](https://github.com/polo-blue/sds/commit/f1ae7145d86a7b4df1976ba850650fc390624667))
+
+## [1.1.6](https://github.com/polo-blue/sds/compare/v1.1.5...v1.1.6) (2025-09-23)
+
+### Bug Fixes
+
+* apply prettier formatting to resolve code style issues ([d9247d2](https://github.com/polo-blue/sds/commit/d9247d2b8809f87ea35777329ba5146b308d809b))
+
+## [1.1.5](https://github.com/polo-blue/sds/compare/v1.1.4...v1.1.5) (2025-09-23)
+
+### Bug Fixes
+
+* resolve additional ESLint warnings and improve type safety ([87ab770](https://github.com/polo-blue/sds/commit/87ab770b0c2e6c33a5936ec693b59e859481857e))
+
 ## [1.1.4](https://github.com/polo-blue/sds/compare/v1.1.3...v1.1.4) (2025-09-23)
 
 ### Bug Fixes